Fix eagle pdf creation for eagle 8
For whatever reason, it fails to run the BRD commands right if they are
in the same script as SCH commands. So we just split them into two
scripts now and run each separately...

Also, if we don't put `quit` as a BRD command, it tries to close eagle
before the actual BRD commands run... so we'll just change that too.
Something definitely changed about scripts in eagle.